HHH-18847 - Organize the org.hibernate.query.results package
