+++ hibernate-core/src/main/java/org/hibernate/generator/values/internal/GeneratedValuesMappingProducer.java
@@ -57,7 +57,6 @@ public class GeneratedValuesMappingProducer implements JdbcValuesMappingProducer
 			final DomainResult<?> domainResult = resultBuilder.buildResult(
 					jdbcResultsMetadata,
 					domainResults.size(),
-					creationState.getLegacyFetchResolver()::resolve,
 					creationState
 			);
