+++ hibernate-core/src/main/java/org/hibernate/procedure/internal/ScalarDomainResultBuilder.java
@@ -4,12 +4,9 @@
  */
 package org.hibernate.procedure.internal;
 
-import java.util.function.BiFunction;
-
 import org.hibernate.metamodel.mapping.BasicValuedMapping;
 import org.hibernate.query.results.ResultBuilder;
 import org.hibernate.query.results.internal.ResultSetMappingSqlSelection;
-import org.hibernate.query.results.internal.dynamic.DynamicFetchBuilderLegacy;
 import org.hibernate.sql.ast.spi.SqlExpressionResolver;
 import org.hibernate.sql.ast.spi.SqlSelection;
 import org.hibernate.sql.results.graph.DomainResult;
@@ -38,7 +35,6 @@ public class ScalarDomainResultBuilder<T> implements ResultBuilder {
 	public DomainResult<T> buildResult(
 			JdbcValuesMetadata jdbcResultsMetadata,
 			int resultPosition,
-			BiFunction<String, String, DynamicFetchBuilderLegacy> legacyFetchResolver,
 			DomainResultCreationState domainResultCreationState) {
 		final SqlExpressionResolver sqlExpressionResolver = domainResultCreationState.getSqlAstCreationState()
 				.getSqlExpressionResolver();
