+++ hibernate-core/src/main/java/org/hibernate/query/results/ResultBuilder.java
@@ -4,14 +4,12 @@
  */
 package org.hibernate.query.results;
 
-import java.util.function.BiConsumer;
-import java.util.function.BiFunction;
-
 import org.hibernate.Incubating;
-import org.hibernate.query.results.internal.dynamic.DynamicFetchBuilderLegacy;
 import org.hibernate.sql.results.graph.DomainResult;
 import org.hibernate.sql.results.graph.DomainResultCreationState;
 import org.hibernate.sql.results.jdbc.spi.JdbcValuesMetadata;
+
+import java.util.function.BiConsumer;
 
 /**
  * Responsible for building a single {@link DomainResult} instance as part of
@@ -26,13 +24,11 @@ public interface ResultBuilder {
 	 *
 	 * @param jdbcResultsMetadata The JDBC values and metadata
 	 * @param resultPosition The position in the domain results for the result to be built
-	 * @param legacyFetchResolver Support for allowing some legacy-style fetch resolution
 	 * @param domainResultCreationState Access to useful stuff
 	 */
 	DomainResult<?> buildResult(
 			JdbcValuesMetadata jdbcResultsMetadata,
 			int resultPosition,
-			BiFunction<String, String, DynamicFetchBuilderLegacy> legacyFetchResolver,
 			DomainResultCreationState domainResultCreationState);
 
 	/**
