+++ hibernate-core/src/main/java/org/hibernate/query/results/ResultBuilderEntityValued.java
@@ -4,9 +4,6 @@
  */
 package org.hibernate.query.results;
 
-import java.util.function.BiFunction;
-
-import org.hibernate.query.results.internal.dynamic.DynamicFetchBuilderLegacy;
 import org.hibernate.sql.results.graph.DomainResultCreationState;
 import org.hibernate.sql.results.graph.entity.EntityResult;
 import org.hibernate.sql.results.jdbc.spi.JdbcValuesMetadata;
@@ -23,6 +20,5 @@ public interface ResultBuilderEntityValued extends ResultBuilder {
 	EntityResult buildResult(
 			JdbcValuesMetadata jdbcResultsMetadata,
 			int resultPosition,
-			BiFunction<String, String, DynamicFetchBuilderLegacy> legacyFetchResolver,
 			DomainResultCreationState domainResultCreationState);
 }
