+++ hibernate-core/src/main/java/org/hibernate/query/results/internal/ResultSetMappingImpl.java
@@ -224,7 +224,6 @@ public class ResultSetMappingImpl implements ResultSetMapping {
 				domainResult = resultBuilder.buildResult(
 						jdbcResultsMetadata,
 						domainResults.size(),
-						creationState.getLegacyFetchResolver()::resolve,
 						creationState
 				);
 			}
